+++ next/src/pages/[name]/index.tsx
@@ -5,8 +5,9 @@ import { useRouter } from 'next/router'
 const UsersIndex: NextPage = () => {
   const router = useRouter()
   const { name } = router.query
+  const authorName = typeof name === 'string' ? name : undefined
 
-  return <Typography>{name}</Typography>
+  return <Typography>{authorName}</Typography>
 }
 
 export default UsersIndex
